Fixed apodization function
In the data flow within data_processing_offline, i. e. process_ft2dir_data -> apodization_function, the info of which function to use was dropped. When apodization_function is called without this parameter, no function is applied and an array of ones is returned silently. So technically the processing works, but the parameter for the apodization/window function has no effect. This has been fixed, and another such function (cos) has been implemented.

This fix brought to light another small bug regarding the window functions that are pre-defined in scipy. There are meant for a full, symmetric interferogramm, where we have only one half. So the window function had to be adjusted.